+++ CHANGELOG.md
@@ -91,6 +91,12 @@ Note: Can be used with `megalinter/megalinter@insiders` in your GitHub Action me
   - [tekton-lint](https://github.com/IBM/tekton-lint) from 0.5.2 to **0.6.0** on 2021-10-21
   - [checkov](https://www.checkov.io/) from 2.0.492 to **2.0.497** on 2021-10-21
   - [terragrunt](https://terragrunt.gruntwork.io) from 0.35.3 to **0.35.4** on 2021-10-21
+  - [stylelint](https://stylelint.io) from 13.13.1 to **14.0.0** on 2021-10-24
+  - [clippy](https://github.com/rust-lang/rust-clippy) from 0.1.55 to **0.1.56** on 2021-10-24
+  - [snakemake](https://snakemake.readthedocs.io/en/stable/) from 6.9.1 to **6.10.0** on 2021-10-24
+  - [sqlfluff](https://www.sqlfluff.com/) from 0.7.0 to **0.7.1** on 2021-10-24
+  - [checkov](https://www.checkov.io/) from 2.0.497 to **2.0.509** on 2021-10-24
+  - [terrascan](https://www.accurics.com/products/terrascan/) from 1.11.0 to **1.12.0** on 2021-10-24
 <!-- linter-versions-end -->
 
 ## [4.46.0] - 2021-09-21