+++ CHANGELOG.md
@@ -98,6 +98,8 @@ Note: Can be used with `nvuillam/mega-linter@insiders` in your GitHub Action meg
   - [checkov](https://www.checkov.io/) from 2.0.507 to **2.0.509** on 2021-10-24
   - [checkov](https://www.checkov.io/) from 2.0.509 to **2.0.510** on 2021-10-25
   - [checkov](https://www.checkov.io/) from 2.0.510 to **2.0.516** on 2021-10-26
+  - [stylelint](https://stylelint.io) from 14.0.0 to **14.0.1** on 2021-10-26
+  - [checkov](https://www.checkov.io/) from 2.0.516 to **2.0.524** on 2021-10-26
 <!-- linter-versions-end -->
 
 ## [4.46.0] - 2021-09-21