+++ CHANGELOG.md
@@ -102,6 +102,9 @@ Note: Can be used with `nvuillam/mega-linter@insiders` in your GitHub Action meg
   - [checkov](https://www.checkov.io/) from 2.0.516 to **2.0.524** on 2021-10-26
   - [php](https://www.php.net) from 7.4.24 to **7.4.25** on 2021-10-27
   - [cfn-lint](https://github.com/martysweet/cfn-lint) from 0.54.3 to **0.54.4** on 2021-10-28
+  - [htmlhint](https://htmlhint.com/) from 0.15.2 to **0.16.0** on 2021-10-29
+  - [terraform-fmt](https://www.terraform.io/docs/cli/commands/fmt.html) from 1.0.9 to **1.0.10** on 2021-10-29
+  - [terragrunt](https://terragrunt.gruntwork.io) from 0.35.4 to **0.35.5** on 2021-10-29
 <!-- linter-versions-end -->
 
 ## [4.46.0] - 2021-09-21